requirements.txt
asgiref==3.3.1
astroid==2.4.2
decouple==0.0.7
dj==0.1
dj-database-url==0.5.0
dj-static==0.0.6
-e git://github.com/django/django@83891e010e37fae8df96a3814c4fcc588586a790#egg=Django
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
wrapt==1.12.1
